+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * `MaxMind::GeoIP2::Client` now validates the IP address before making a
   request to the web service.
+* `MaxMind::GeoIP2::Client` now includes the version of Ruby, the version
+  of the HTTP client library, and its own version in the User-Agent header.
 
 ## 1.1.0 (2021-11-18)